+++ stories/components/checkbox/checkbox.stories.ts
@@ -12,7 +12,9 @@ const meta: Meta = {
     parameters: {
         actions: {
             handles: [
-                'wcsChange'
+                'wcsBlur',
+                'wcsChange',
+                'wcsFocus'
             ]
         }
     },