+++ CHANGELOG.md
@@ -13,6 +13,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - **select**: add a red outline when the `wcs-form-field` is in error
 - **native-select**: add a red outline when the `wcs-form-field` is in error
+- **checkbox**:
+    - add wcsFocus and wcsBlur events
+    - add onTouched support for Angular
 
 ### Changed